added Pseudo soln for Potts, bug fixes, and minor updates Pseudo can now be used to solve k-state Potts model, tested out for k=3 fixed models.Potts3 bug in setting calc_e update models.Model to handle Potts3 model fixed bugs in model.Model sampling methods added utils.define_pseudo_potts_helper_functions() bug in setting custom Pseudo renamed define_potts3_helper_functions() to define_potts_helper_functions() and generalized it to arbitrary k

+++ coniii/samplers.py
@@ -1419,7 +1419,7 @@ class Metropolis(Sampler):
             def f(args):
                 s, E, nSamples, seed = args
                 rng = np.random.RandomState(seed)
-                samples = np.zeros((nSamples, self.n), dtype=int)
+                samples = np.zeros((nSamples, self.n), dtype=np.int8)
 
                 for i in range(burn_in):
                     de = self.sample_metropolis( s, E, rng=rng )
@@ -1435,7 +1435,7 @@ class Metropolis(Sampler):
             def f(args):
                 s, E, nSamples, seed = args
                 rng = np.random.RandomState(seed)
-                samples = np.zeros((nSamples, self.n), dtype=int)
+                samples = np.zeros((nSamples, self.n), dtype=np.int8)
 
                 for i in range(burn_in):
                     de = self.sample_metropolis( s, E, rng=rng, flip_site=i%self.n )
